fix: pytest

+++ tests/indexes/test_indexes.py
@@ -1531,7 +1531,7 @@ def test_array_property() -> None:
             pd.Categorical,
         ),
         pd.Categorical,
-        np.int64,
+        int,
     )
     check(
         assert_type(pd.interval_range(0, 1).array, IntervalArray),
